defect #1123479 [Jira Plugin] Test connection does not detect that the location base is already used when creating a new space
Problem: when in the dialog of adding a new space configuration, you can test connection before saving. If the connection is successful, a green icon is shown, but if you press save and the name/location is not valid, the icon will turn red.

Solution: validate the uniqueness of space name/location on test connection, mentioning in the message if the connectivity can be established but other validations fail.